+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/security_evaluator/AppliedNetworkActionsResult.java
@@ -1,4 +1,4 @@
-package com.farao_community.farao.swe_csa.app.dichotomy;
+package com.farao_community.farao.swe_csa.app.security_evaluator;
 
 import com.powsybl.iidm.network.Country;
 import com.powsybl.openrao.data.crac.api.RemedialAction;
